requirements.txt
flask==2.3.3
twilio==8.10.0
google-genai>=0.6.0
python-dotenv==1.0.0
requests==2.31.0
psutil==5.9.6
pydantic==2.5.0
